+++ types/index.d.ts
@@ -1,5 +1,6 @@
 export interface EngineOptions {
-  allowUndefinedFacts: boolean;
+  allowUndefinedFacts?: boolean;
+  pathResolver?: PathResolver;
 }
 
 export interface EngineResult {
@@ -92,6 +93,11 @@ export interface Event {
   type: string;
   params?: Record<string, any>;
 }
+
+export type PathResolver = (
+  value: object,
+  path: string,
+) => any;
 
 export type EventHandler = (
   event: Event,