+++ CHANGELOG.md
@@ -1,15 +1,19 @@
 #### 6.0. / 2020-12-XX
   * BREAKING CHANGES
-    * `path` support using `selectn` should use the `pathResolver` feature. Read more [here](). To continue using selectn, add the following to the engine constructor:
+    * `path` support using `selectn` should use the `pathResolver` feature. Read more [here](./docs/rules.md#condition-helpers-custom-path-resolver). To continue using `selectn`, add the following to the engine constructor:
       ```js
       const pathResolver = (object, path) => {
         return selectn(path)(object)
       }
       const engine = new Engine(rules, { pathResolver })
       ```
+      (fixes #205)
     * Engine and Rule events `on('success')`, `on('failure')`, and Rule callbacks `onSuccess` and `onFailure` now honor returned promises; any event handler that returns a promise will be waited upon to resolve before engine execution continues. (fixes #235)
     * Private `rule.event` property renamed. Use `rule.getEvent()` to avoid breaking changes in the future.
     * The 'success-events' fact used to store successful events has been converted to an internal data structure and will no longer appear in the almanac's facts. (fixes #187)
+  * NEW FEATURES
+    * Engine constructor now accepts a `pathResolver`   option for resolving condition `path` properties. Read more [here](./docs/rules.md#condition-helpers-custom-path-resolver). (fixes #210)
+
 
 #### 5.3.0 / 2020-12-02
   * Allow facts to have a value of `undefined`