+++ CHANGELOG.md
@@ -13,7 +13,7 @@
     * The `success-events` fact used to store successful events has been converted to an internal data structure and will no longer appear in the almanac's facts. (fixes #187)
   * NEW FEATURES
     * Engine constructor now accepts a `pathResolver` option for resolving condition `path` properties. Read more [here](./docs/rules.md#condition-helpers-custom-path-resolver). (fixes #210)
-    * Engine.run() now returns:
+    * Engine.run() now returns three additional data structures:
       * `failureEvents`, an array of all failed rules events. (fixes #192)
       * `results`, an array of RuleResults for each successful rule (fixes #216)
       * `failureResults`, an array of RuleResults for each failed rule